requirements.txt
PyQt5~=5.15.7
pyqt5-tools~=5.12
PyQt5-stubs~=5.15
pyinstaller~=5.2
PyYAML~=6.0
requests~=2.28.0
openpyxl~=3.0.7
psutil~=5.8.0
pyserial~=3.5
setuptools~=63.2.0